visp_mimo : Fix mimo mbox crash druing mimo_video probe.
Mbox crash observed during mimo_video probe.
This crash is observed after 32bit variabled are converted to
64-bit in visp/include/media_device.h

+++ visp_mimo/include/media_device.h
@@ -229,7 +229,7 @@ typedef struct {
 } media_fmt;
 
 typedef struct {
-	uint32_t dma_addr;
+	uint64_t dma_addr;
 	uint32_t dma_size;
 } media_plane;
